requirements.txt
torch==2.1.2
transformers==4.42.3
datasets==2.16.1
haystack-ai==2.1.2
qdrant-haystack
evaluate==0.4.1
peft==0.11.1
trl==0.8.6
bitsandbytes==0.42.0
sentencepiece==0.2.0
sentence_transformers==2.4.0
deepeval==0.21.73
bert_score==0.3.13

# Extra Tools
pyyaml==6.0.1
nltk==3.8.1
rouge_score==0.1.2
wandb==0.16.4
hydra-core==1.3.2
pre-commit==3.6.0
openai==1.23.3
